Update launch-a-basic-appium-2-session.adoc
Kris J determined that devices with passcodes are not supported with basic Appium 2. The problem is that when a Basic Appium session is started, deviceConnect stops our running agent and starts the Appium WebDriverAgent, and that WebDriverAgent process often asks that the XCTest passcode to be re-entered, which we can't do automatically. Looking into a fix but adding this for now.